Configure DatabasePager threads correctly
Previously we didn't set the number of DatabasePager
threads on initialization, and so irrespective of
the /sim/rendering/database-pager/threads property
we always had 1 General thread and 1 HTTP thread.

(On reset, we did have some code that set up the correct
number of threads).

This wasn't completely obvious, as the threads are configured
at startup using an OSG DisplaySettings hint.

This change sets up these threads correctly on initialization
and ties the properties properly.  Though it's worth noting
that changing the property in-sim will have no effect until a
Reset.