Add $FG_OVERRIDE_PAGER_THREADS to set load threads
Previously /sim/rendering/database-pager/threads was used to set
the number of OSG DatabasePager threads one reset (only!).

This change removes this function and replaces it with an
environmental variable FG_OVERRIDE_PAGER_THREADS which will
set the number of threads at startup and on reset.

This ensures that only developers who are interested in
multi-threading will enable it.

It is expected that one multi-threaded DatabasePagers stabilizes
we will change to use /sim/rendering/database-pager/threads
instead (for both init an reset).