Canvas: Fix placements not working for glTF models
Canvas was using the default state set provided by an Effect.
This does not work if the Effect doesn't have a fixed-function
pipeline technique (like in the case of HDR and model-pbr).

In the future we should remove placements by texture name entirely
and use the node name together with a texture type (base color,
metallic, roughness) to identify the correct texture unit.