WS30: Add support for compressed (zip) tiles
See also simgear
4400ab164fc55493b9b487c3680baa545e0cba65

This commit add support for zipped osgdem output.

This leverages OSG's built in support for compressed archives,
and specifically zip.

It includes a workaround to a bug in the OSG path handling for
a hierarchy of osg files by assuming that all zip files do not
contain directories.  (This is the "-j" zip option).