input: Improve code readability and debug output
This commit renames a few variables to more speaking names and adds
comments.
Some debug output has been added or modified for better
understanding of code flow.
A class_id was added and is exported to the property tree to indicate
which FG input system has created which device in /input/devices