Fix ground vehicle tunnel elevation
Previously the start and end elevation of tunnels
were stored as static variables.  So all ground
vehicles inside tunnels were assumed to be in the same
tunnel.  This was probably find due to the very limited
number of train AI.

Now the tunnel start and end elevation is stored as a
member variable on the AI object.